+++ AircallLaurent/AircallLaurent/SupportingFiles/Assets.xcassets/AppIcon.appiconset/Contents.json
